Restructure, extracting forms
